fix(auth): correct role routes authorization function for resources
