+++ src/routes/user.routes.js
@@ -1,7 +1,8 @@
 // routes/user.routes.js
 import express from "express";
 import userController from "../controllers/user.controller.js";
-import { authenticateJWT, authorize } from "../middlewares/auth.middleware.js";
+import { authenticateJWT} from "../middlewares/auth.middleware.js";
+import { authorize } from "../middlewares/role.middleware.js";
 import { updateUserValidator } from "../middlewares/validators/user.validator.js";
 
 const router = express.Router();
